+++ Graphics/Include/DenOfIzGraphics/Assets/Serde/Mesh/MeshAsset.h
@@ -76,7 +76,7 @@ namespace DenOfIz
     struct DZ_API MeshVertexArray
     {
         MeshVertex *Elements;
-        uint32_t    NumElements;
+        size_t      NumElements;
     };
 
     struct DZ_API MorphTargetDeltaAttributes
